+++ src/chapters/02/test/ex10.test.ts
@@ -1,0 +1,13 @@
+// See [Explicit URL Imports](https://vitejs.dev/guide/assets.html#explicit-url-imports)
+import popularNames from '@/assets/popular-names.txt'
+import { execSync } from 'node:child_process'
+import { ex10 } from '@/chapters/02/ex10'
+
+describe('10. Line count', () => {
+  it('Confirm the result by using `wc` command', async () => {
+    // $ wc -l < popular-names.txt
+    const wc = execSync(`wc -l < .${popularNames}`).toString()
+
+    expect(await ex10()).toBe(Number(wc))
+  })
+})